+++ src/io/mod.rs
@@ -1,4 +1,4 @@
 //! Our own format readers and writers. Zero dependencies.
 
 pub mod csv;
-// pub mod json;
+pub mod json;